write test for getting user name from local storage
