+++ src/scenes/UserInfo.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-alert */
 import Phaser from 'phaser';
-import { createUserInputForm } from '../utils/common';
+import createUserInputForm from '../utils/user';
 
 export default class BootScene extends Phaser.Scene {
   constructor() {
